update flag RESET_DATAa

+++ docker-entrypoint.sh
@@ -42,17 +42,18 @@ if [ "$1" = 'redis-cluster' ]; then
     for port in $(seq $INITIAL_PORT $max_port); do
       mkdir -p /redis-conf/${port}
       mkdir -p /redis-data/${port}
+      if [ -z "$RESET_DATA" -o "$RESET_DATA" = "true" ]; then
+        if [ -e /redis-data/${port}/nodes.conf ]; then
+          rm /redis-data/${port}/nodes.conf
+        fi
 
-      if [ -e /redis-data/${port}/nodes.conf ]; then
-        rm /redis-data/${port}/nodes.conf
-      fi
+        if [ -e /redis-data/${port}/dump.rdb ]; then
+          rm /redis-data/${port}/dump.rdb
+        fi
 
-      if [ -e /redis-data/${port}/dump.rdb ]; then
-        rm /redis-data/${port}/dump.rdb
-      fi
-
-      if [ -e /redis-data/${port}/appendonly.aof ]; then
-        rm /redis-data/${port}/appendonly.aof
+        if [ -e /redis-data/${port}/appendonly.aof ]; then
+          rm /redis-data/${port}/appendonly.aof
+        fi
       fi
       
       if [ -z "$PROTECTED_MODE" -o "$PROTECTED_MODE" = "true" ]; then
@@ -94,25 +95,26 @@ if [ "$1" = 'redis-cluster' ]; then
     ## If it is below 5.0 then we use the redis-trib.rb to build the cluster
     #
     /redis/src/redis-cli --version | grep -E "redis-cli 3.0|redis-cli 3.2|redis-cli 4.0"
+    if [ -z "$RESET_DATA" -o "$RESET_DATA" = "true" ]; then
+        if [ $? -eq 0 ]
+        then
+          echo "Using old redis-trib.rb to create the cluster"
+          echo "yes" | eval ruby /redis/src/redis-trib.rb create --replicas "$SLAVES_PER_MASTER" "$nodes"
+        else
+          echo "Using redis-cli to create the cluster"
+          if [ -z "$PASSWORD"  ]; then
+            echo "yes" | eval /redis/src/redis-cli --cluster create --cluster-replicas "$SLAVES_PER_MASTER" "$nodes"
+            password_arg="-a $PASSWORD"
+          else
+            echo "yes" | eval /redis/src/redis-cli --cluster create --cluster-replicas "$SLAVES_PER_MASTER" -a "$PASSWORD" "$nodes"
+          fi
+        fi
 
-    if [ $? -eq 0 ]
-    then
-      echo "Using old redis-trib.rb to create the cluster"
-      echo "yes" | eval ruby /redis/src/redis-trib.rb create --replicas "$SLAVES_PER_MASTER" "$nodes"
-    else
-      echo "Using redis-cli to create the cluster"
-      if [ -z "$PASSWORD"  ]; then
-        echo "yes" | eval /redis/src/redis-cli --cluster create --cluster-replicas "$SLAVES_PER_MASTER" "$nodes"
-        password_arg="-a $PASSWORD"
-      else
-        echo "yes" | eval /redis/src/redis-cli --cluster create --cluster-replicas "$SLAVES_PER_MASTER" -a "$PASSWORD" "$nodes"
-      fi
-    fi
-
-    if [ "$SENTINEL" = "true" ]; then
-      for port in $(seq $INITIAL_PORT $(($INITIAL_PORT + $MASTERS))); do
-        redis-sentinel /redis-conf/sentinel-${port}.conf &
-      done
+        if [ "$SENTINEL" = "true" ]; then
+          for port in $(seq $INITIAL_PORT $(($INITIAL_PORT + $MASTERS))); do
+            redis-sentinel /redis-conf/sentinel-${port}.conf &
+          done
+        fi
     fi
 
     tail -f /var/log/supervisor/redis*.log
